# docs/signaling-openapi.yaml
openapi: 3.0.2
servers:
  - url: https://example.com/signaling/v1
info:
  description: |-
    This specification contains all mandatory and optional endpoints for the [Dataplane Signaling Protocol (DPS)](https://github.com/Metaform/dataplane-signaling/blob/main/docs/signaling.md).
    It is based on OpenAPI 3.0.2 and uses JSON Schema for data modeling.
  version: 0.0.1
  title: Dataplane Signaling Protocol DPS
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
...
paths:
...
  /dataflows/{id}/resume:
    post:
      tags:
        - Data plane endpoints
      summary: Resume a suspended DataFlow
      description: |
        The resume request signals to the data plane to synchronously resume a suspended data transfer.
        Both the request and response may carry a dataAddress so updated endpoint information can be propagated
        during suspend/start cycles.
      operationId: resumeDataFlowById
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DataFlowResumeMessage'
      parameters:
        - name: id
          in: path
          required: true
          description: The unique identifier of the data flow to resume
          schema:
            type: string
            example: flow-1234
      responses:
        '200':
          description: DataFlow resumed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransferStartResponseMessage'
              example:
                dataAddress:
                  '@type': DataAddress
                  endpointType: https://23id.org.idsa/v4.1/HTTP
                  endpoint: https://example.com/api/data
                  endpointProperties:
                    - '@type': EndpointProperty
                      name: authorization
                      value: 5up3r53cur3t0k3n
                    - '@type': EndpointProperty
                      name: authType
                      value: bearer
        '400':
          description: Bad Request - invalid input, invalid object state or missing required parameters
        '404':
          description: Not Found - data flow with specified ID does not exist
...
components:
  schemas:
...
    DataFlowResumeMessage:
      type: object
      required: [ messageId, processId ]
      properties:
        messageId:
          type: string
          description: Unique identifier for the resume message
          example: b1d5f9e2-3c4b-4f7a-9c3e-2f1e5d6c7b8a
        processId:
          type: string
          description: The transfer process ID as assigned by the control plane for correlation
          example: test-transfer-process-id
        dataAddress:
          $ref: '#/components/schemas/DataAddress'
    TransferStartResponseMessage:
      type: object
      description: |
        Response body used to propagate endpoint updates during transfer start/resume handling.
        The containing object allows adding fields in future versions.
      properties:
        dataAddress:
          $ref: '#/components/schemas/DataAddress'
...
    DataAddress:
      type: object
      required:
        - type
        - endpointType
        - endpoint
        - endpointProperties
      properties:
        '@type':
          type: string
          description: The type of data address
          example: DataAddress
        endpointType:
          type: string
          description: the type of endpoint, indicating the protocol to be used
          example: https://23id.org.idsa/v4.1/HTTP
        endpoint:
          type: string
          description: The network location of the endpoint
          example: https://example.com/api/data
          format: uri
        endpointProperties:
          type: array
          description: Additional properties for the endpoint
          items:
            $ref: '#/components/schemas/EndpointProperty'
          example:
            '@type': DataAddress
            endpointType: https://w3.org/idsa/v4.1/HTTP
            endpoint: https://example.com/api/data
            endpointProperties:
              - '@type': EndpointProperty
                name: authorization
                value: 5up3r53cur3t0k3n
              - '@type': EndpointProperty
                name: authType
                value: bearer
    EndpointProperty:
      type: object
      required:
        - '@type'
        - name
        - value
      properties:
        '@type':
          type: string
          default: EndpointProperty
        name:
          type: string
          description: The name of the property
          example: authorization
        value:
          type: string
          description: The value of the property
          example: 5up3r53cur3t0k3n
...
  securitySchemes:
    petstore_auth:
      type: oauth2
      flows:
        implicit:
          authorizationUrl: 'https://petstore.swagger.io/oauth/authorize'
          scopes:
            'write:pets': modify pets in your account
            'read:pets': read your pets
    api_key:
      type: apiKey
      name: api_key
      in: header